requirements.txt
Django==3.2.11
dj-database-url==0.5.0
psycopg2-binary==2.9.3
python-decouple==3.6
drf-nested-routers==0.93.4
djangorestframework==3.13.1